move stuff from utils to html_processors.py
