+++ niworkflows/info.py
@@ -50,7 +50,7 @@ CLASSIFIERS = [
 ]
 
 REQUIRES = ['nipype', 'future', 'nilearn>=0.2.6', 'sklearn', 'pandas', 'matplotlib',
-            'jinja2', 'svgutils', 'tinycss']
+            'jinja2', 'svgutils', 'tinycss', 'beautifulsoup4']
 SETUP_REQUIRES = []
 REQUIRES += SETUP_REQUIRES
